Greatly refactor dependency resolution
A new method is introduced that can properly model the duplicate
occurrences of bundles if they are part of multiple resolution domains.
The callers should manually consume the direct dependencies of each
bundle rather than working with them as a whole.

The old API is still kept in place for compatibility, and tests pass for
them, but the private dependencies won't resolve using the old API.

Further tests still yet to be added as some parts of the repository
still use the old API and classloaders also need some modifications.