Refactor the dependency resolution some more
The algorithm for tracking the pinned bundles and backtracking have been
modified so it is easier to manage and can properly support private
dependencies.

The performance of the dependency resolution can possibly be improved,
however, it should work correctly as this implementation is cleaner than
the previous.

More tests could still be added.