Implement private dependencies (#6)
* Add private meta-data constant and query method

* Base implementation of private dependencices

There should be more unit and integration tests added. Especially class
loading related.

* Greatly refactor dependency resolution

A new method is introduced that can properly model the duplicate
occurrences of bundles if they are part of multiple resolution domains.
The callers should manually consume the direct dependencies of each
bundle rather than working with them as a whole.

The old API is still kept in place for compatibility, and tests pass for
them, but the private dependencies won't resolve using the old API.

Further tests still yet to be added as some parts of the repository
still use the old API and classloaders also need some modifications.

* Add deprecations

* Refactor class loading to use new dependency resolution API

* Simple private dependency classloader test

We needed to change the classloader implementation to track the loaded
classes instead of using findLoadedClass method as that may return
classes loaded by other bundles.

* Common and circular classloader private dependency tests

* Test classloader resolution of already published bundles

Test for loading known bundles and examining their classes to ensure
that they can be properly loaded with the new dependency resolution
algorithm. The test checks that we don't break anything that is working
with the new version.

* Refactor the dependency resolution some more

The algorithm for tracking the pinned bundles and backtracking have been
modified so it is easier to manage and can properly support private
dependencies.

The performance of the dependency resolution can possibly be improved,
however, it should work correctly as this implementation is cleaner than
the previous.

More tests could still be added.

* Overlapping via private dependencies tests

* JavaDoc for the added or modified source elements