Fix class loader identifier usage
The determination of class loader identifiers for serialization were
wrong. They incorrectly created the classloaders with incorrect
dependency resolution and configuration.

We now generate the classloader identifiers based on the domain of the
classloaders and their reachable dependencies. Now the classloaders can
be properly reconstructed based on the identifiers.

We use JSON objects to generate the identifiers, however, this can
change in the future for a proprietary format. (Actually we should test
if that is faster or not. Or smaller.)