Support loading the same native library into multiple classloaders
Previously if a native library was already loaded in another
classloader, we would fail the loading with an UnsatisfiedLinkError.
With this enhancement we keep track of the loaded libraries via file
locks so it works between different classloaders.

This allows loading a native library multiple times so reloading the
saker.build runtime won't cause errors if the old classloaders haven't
been garbage collected yet.