Handle possible exception in LazyReflectionTaskInformation
Exceptions can occurr if the class cannot be loaded for whatever
reasons. E.g. dependencies cannot be satisfied.
Instead of having it propagate, just return null signaling that the
associated information cannot be retrieved. As this is an IDE modelling
feature, this is better than having an exception stacktrace displayed to
the user.
It is recommended that we somehow handle the exception better in the
future.